Changed the page classes of put patch delete
